requirements.txt
# requirements.txt
# Core dependencies
fastapi==0.68.0
uvicorn==0.15.0
jinja2==3.0.1
aiofiles==0.7.0
python-multipart==0.0.5
python-dotenv==0.19.0
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
bitsandbytes==0.41.0
sentencepiece==0.1.99
protobuf==4.25.1
pydantic==1.10.0

# Security dependencies
python-jose==3.3.0
cryptography==41.0.1
pycryptodome==3.18.0
pyOpenSSL==23.2.0
pyjwt==2.6.0

# Monitoring and metrics
prometheus-client==0.16.0

# GCP dependencies
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4
google-cloud-monitoring==2.15.1
google-cloud-logging==3.8.0

# Additional utilities
pyotp==2.8.0
qrcode==7.3

# Add these lines
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
